fixes for sed

+++ distros/centos/7/minimal_install_setup.sh
@@ -10,17 +10,19 @@ packages="vim screen net-tools"
 # fixup eno16777736 -> eth0
 eth_name=eth0
 eth_file=/etc/sysconfig/network-scripts/ifcfg-${eth_name}
-mv /etc/sysconfig/network-scripts/ifcfg-eno16777736  ${eth_file}
+if [ -f /etc/sysconfig/network-scripts/ifcfg-eno16777736 ]; then 
+  mv /etc/sysconfig/network-scripts/ifcfg-eno16777736  ${eth_file}
+fi
 
 # fixup grub to match
-sed -i.bak 's/.*GRUB_CMDLINE_LINUX=.*/GRUB_CMDLINE_LINUX="rd.lvm.lv=centos/swap vconsole.font=latarcyrheb-sun16 rd.lvm.lv=centos/root crashkernel=auto  vconsole.keymap=us rhgb quiet net.ifnames=0 biosdevname=0"/' /etc/sysconfig/grub
+sed -i'' 's/.*GRUB_CMDLINE_LINUX=.*/GRUB_CMDLINE_LINUX="rd.lvm.lv=centos\/swap vconsole.font=latarcyrheb-sun16 rd.lvm.lv=centos\/root crashkernel=auto  vconsole.keymap=us rhgb quiet net.ifnames=0 biosdevname=0"/' /etc/sysconfig/grub
 grub2-mkconfig  -o /boot/grub2/grub.cfg
 
 # Enable network, wipe device and UUID, change name to eth0
-sed -i.bak "s/NAME=.*/NAME=${eth_name}/" ${eth_file}
-sed -i.bak 's/ONBOOT=.*/ONBOOT=yes/' ${eth_file}
-sed -i.bak '/UUID=.*/d' ${eth_file}
-sed -i.bak '/DEVICE=.*/d' ${eth_file}
+sed -i'' "s/NAME=.*/NAME=${eth_name}/" ${eth_file}
+sed -i'' 's/ONBOOT=.*/ONBOOT=yes/' ${eth_file}
+sed -i'' '/UUID=.*/d' ${eth_file}
+sed -i'' '/DEVICE=.*/d' ${eth_file}
 
 
 # Disable ipv6
@@ -28,7 +30,7 @@ echo 'net.ipv6.conf.all.disable_ipv6 = 1' > /etc/sysctl.d/disableipv6.conf
 sysctl -p
 
 # Disable selinux
-sed -i.bak 's/SELINUX=.*/SELINUX=permissive/g' /etc/sysconfig/selinux
+sed -i'' 's/SELINUX=.*/SELINUX=permissive/g' /etc/sysconfig/selinux
 setenforce 0
 
 # disable firewalld
@@ -51,7 +53,7 @@ me=`basename "$0"`
 rm ${me}
 
 # clear history
-history -c
+> /root/.bash_history
 
 # shutdown and restart
 shutdown -r now
